+++ backend/tests/MyProject/src/data-source.ts
@@ -1,12 +1,12 @@
 import "reflect-metadata"
 import { DataSource } from "typeorm"
-import { Project } from "./entity/project"
-import { User } from "./entity/user"
-import { Release } from "./entity/release"
-import { Revision } from "./entity/revision"
-import { Sprint } from "./entity/sprint"
-import { Story } from "./entity/story"
-import { Task } from "./entity/task"
+import { Project } from "../../../src/entity/project"
+import { User } from "../../../src/entity/User"
+import { Release } from "../../../src/entity/release"
+import { Revision } from "../../../src/entity/revision"
+import { Sprint } from "../../../src/entity/sprint"
+import { Story } from "../../../src/entity/story"
+import { Task } from "../../../src/entity/task"
 
 export const AppDataSource = new DataSource({
     type: "postgres",
